Support title replace in addition to prefix
The approach is backward compatible with existing workflows -- if the
new title-replace-template setting is not configured the title is
prefixed as before.

The set/prefix can be used mutually exclusively, or additively. The
former is able to set or prefix the title. The latter can set the main
part of the title, and then prefix it.